Dataset used in ISMIR (2015) paper
This is the dataset, as python pickle files, as used in the following paper:
Janssen, B., Kranenburg, P. V., & Volk, A. (2015). A Comparison of Symbolic Similarity Measures for Finding Occurrences of Melodic Segments. In Proc. 16th International Society for Music Information Retrieval Conference.

It contains 213 melodies and 1084 phrases, as opposed to the 360 melodies and 1891 phrases contained in the MeertensTuneCollections-ANN2.0, as found here:
http://www.liederenbank.nl/mtc/